api/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-storage
google-cloud-video-transcoder
google-cloud-speech
google-genai
pydantic
python-multipart
python-dotenv
slowapi
opencv-python-headless
mediapipe
google-adk
google-cloud-aiplatform
httpx
websockets
